TC-2826: refactored LicenseInfo
