+++ modules/fundamental/src/common/mod.rs
@@ -1,5 +1,8 @@
+use crate::sbom::service::sbom::LicenseBasicInfo;
 use sea_orm::FromQueryResult;
+use sea_query::FromValueTuple;
 use serde::{Deserialize, Serialize};
+use trustify_entity::sbom_package_license::LicenseCategory;
 use utoipa::ToSchema;
 
 pub mod service;
@@ -8,4 +11,19 @@ pub mod service;
 pub struct LicenseRefMapping {
     pub license_id: String,
     pub license_name: String,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, ToSchema)]
+pub struct LicenseInfo {
+    pub license_name: String,
+    pub license_type: LicenseCategory,
+}
+
+impl From<LicenseBasicInfo> for LicenseInfo {
+    fn from(license_basic_info: LicenseBasicInfo) -> Self {
+        LicenseInfo {
+            license_name: license_basic_info.license_name,
+            license_type: LicenseCategory::from_value_tuple(license_basic_info.license_type),
+        }
+    }
 }
